requirements.txt
selene>=2.0.0rc4
pytest~=7.3.1
selenium~=4.10.0
pydantic
pydantic_settings
python-dotenv
allure-pytest
Appium-Python-Client
requests